Muevo test de integración a un módulo aparte
